STM32F401RE-Nucleo/Projects/drivers_development_tutorial/drivers/include/stm32f401xe_driver_i2c.h: type of fast_mode_duty_cycle member of i2c_config_t changed from uint16_t to uint8_t

+++ STM32F401RE-Nucleo/Projects/drivers_development_tutorial/drivers/include/stm32f401xe_driver_i2c.h
@@ -45,7 +45,7 @@ typedef struct {
     uint32_t clock_speed;
     uint8_t device_address;
     uint8_t ack_control;
-    uint16_t fast_mode_duty_cycle;
+    uint8_t fast_mode_duty_cycle;
 }i2c_config_t;
